pyproject.toml
[tool.poetry]
name = "app"
version = "0.2.0"
description = ""
authors = [

]
maintainers = [

]
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.12, <3.13"
fastapi = "^0.111.0"
uvicorn = { version = "^0.30.1", extras = ["standard"] }
pydantic = "^2.7.2"
pydantic-settings = "^2.2.1"
yarl = "^1.9.4"
ujson = "^5.10.0"
tortoise-orm = "^0.21.3"
aerich = "^0.7.2"
asyncpg = "^0.29.0"
httptools = "^0.6.1"
loguru = "^0.7.2"
passlib = "^1.7.4"
python-jose = "^3.3.0"
pytest-fastapi-deps = "^0.2.3"
orjson = "^3.10.3"
pyjwt = "^2.8.0"
boto3 = "^1.34.137"
ptvsd = "^4.3.2"
requests = "^2.32.3"
types-requests = "^2.32.0.20240712"


[tool.poetry.dev-dependencies]
pytest = "^8.2.1"
flake8 = "~7.0.0"
mypy = "^1.10.0"
isort = "^5.13.2"
pre-commit = "^3.7.1"
wemake-python-styleguide = "^0.19.2"
black = "^24.4.2"
autoflake = "^2.3.1"
pytest-cov = "^5.0.0"
anyio = "^4.4.0"
pytest-env = "^1.1.3"
asynctest = "^0.13.0"
nest-asyncio = "^1.6.0"
httpx = "^0.27.0"

[tool.isort]
profile = "black"
multi_line_output = 3
src_paths = ["app",]

[tool.mypy]
strict = true
ignore_missing_imports = true
allow_subclassing_any = true
allow_untyped_calls = true
pretty = true
show_error_codes = true
implicit_reexport = true
allow_untyped_decorators = true
warn_unused_ignores = false
warn_return_any = false
namespace_packages = true

[tool.pytest.ini_options]
filterwarnings = [
    "error",
    "ignore::DeprecationWarning",
    "ignore:.*unclosed.*:ResourceWarning",
]
env = [
    "APP_ENVIRONMENT=pytest",
    "APP_DB_BASE=app_test",
]

[tool.aerich]
tortoise_orm = "app.domains.database.TORTOISE_CONFIG"
location = "./migrations"
src_folder = "./."

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
